+++ src/Validator.js
@@ -1,5 +1,6 @@
 import { computed, action, observable } from 'mobx';
 import _ from 'lodash';
+import { $try } from './parser';
 
 import VJF from './validators/VJF'; // Vanilla JavaScript Functions
 import SVK from './validators/SVK'; // Json Schema Validation Keywords
@@ -52,10 +53,57 @@ export default class Validator {
       })));
   }
 
+  validate(opt = {}, obj = {}) {
+    const form = opt.form;
+    action(() => (form.$validating = true))();
+    action(() => (this.$genericErrorMessage = null))();
+
+    const path = $try(opt.path, opt);
+    const field = $try(opt.field, form.select(path, null, null));
+    const related = $try(opt.related, obj.related, false);
+    const $showErrors = $try(opt.showErrors, obj.showErrors, true);
+
+    form.state.events.set('validate', field ? field.path : true);
+    // look running events and choose when show errors messages
+    const notShowErrorsEvents = ['clear', 'reset'];
+    if (form.state.options.get('showErrorsOnUpdate') === false) notShowErrorsEvents.push('update');
+    const showErrors = $showErrors && !form.state.events.running(notShowErrorsEvents);
+
+    // wait all promises then resolve
+    const $wait = resolve => Promise.all(this.promises)
+      .then(action(() => (form.$validating = false)))
+      .then(() => form.state.events.set('validate', false))
+      .then(() => resolve(form.isValid));
+
+    if (_.isPlainObject(opt) && !_.isString(path)) {
+      // validate all fields
+      return new Promise((resolve) => {
+        this.validateAll({
+          showErrors,
+          related,
+          form,
+        });
+
+        return $wait(resolve);
+      });
+    }
+
+    // validate single field by path
+    return new Promise((resolve) => {
+      this.validateField({
+        showErrors,
+        related,
+        form,
+        field,
+        path,
+      });
+
+      return $wait(resolve);
+    });
+  }
+
   @action
   validateAll({ form, showErrors = true, related = false }) {
-    // reset generic error message
-    this.resetGenericError();
     // validate all fields and nested fields
     form.forEach(field =>
       this.validateField({
@@ -108,11 +156,6 @@ export default class Validator {
     const $default = this.options.get('defaultGenericError');
     if (_.isString($default)) return $default;
     return 'The form is invalid';
-  }
-
-  @action
-  resetGenericError() {
-    this.$genericErrorMessage = null;
   }
 
   @action